% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip}
\alias{strip}
\title{Visualise large amounts of hourly data as yearly heatmaps}
\usage{
strip(x, date, fun = mean, range, cond = rep(" ", length(x)),
  arrange = c("long", "wide"),
  colour = colorRampPalette(rev(brewer.pal(11, "Spectral"))),
  n.col.levs = 1000, start, end, ...)
}
\arguments{
\item{x}{the variable to be plotted.}

\item{date}{the time series variable to be used for plotting. Needs to be 
YYYY-MM-DD HH:MM format}

\item{fun}{the function to be used in case the time series is higher resolution 
than hourly.}

\item{range}{the range of x to be plotted. Values outside this range will not
be plotted.}

\item{cond}{the conditioning variable to be used for creating the various strips.
Usually the year. If not supplied, year will be used by default.}

\item{arrange}{'long' (the default) will render all strips in one column.
'wide' will render things depending on device size.}

\item{colour}{the colour palette to be used for the heatmap.}

\item{n.col.levs}{the number of colour levels to interpolate to. 
Increase this if the colour palette does not span all data values.}

\item{start}{the start date of the time series to plot. Only relevant for short
time periods (i.e. sub-year).}

\item{end}{the end of the time series to plot. Corresponding to \code{start}.}
}
\description{
Produces a heatmap display of hourly time series data showing hour of day on
the y-axis and day-of-year in the x-axis
}
